feat: moved RE's entities
