+++ src/main/java/it/gov/pagopa/wispconverter/technicalsupport/controller/mapper/ReEventMapper.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.wispconverter.technicalsupport.controller.mapper;
 
 import it.gov.pagopa.wispconverter.technicalsupport.controller.model.ReEvent;
-import it.gov.pagopa.wispconverter.technicalsupport.repository.model.ReEventEntity;
+import it.gov.pagopa.wispconverter.technicalsupport.repository.model.re.ReEventEntity;
 import it.gov.pagopa.wispconverter.technicalsupport.util.Constants;
 import org.mapstruct.*;
